Move configuration into separate module
Config needs to be loaded before routes so that the api keys are
available. But the way I had it, the load was happening before another
import, which isn't technically valid es2015. This change puts the
configuration into another module, which is intended to be imporetd for
its side effects. This allows all imports to be grouped together at the
top of the module.